refactor: change namespace of graphql php attributes
